Adjustments and Conversion
Adjustede the grammar of the parser so it uses a vertical bar | instead
of a forward slash for OR.
Moved the .bat file to be in the same folder as the python file it's
accessing. Adjusted code as needed for that to run.
Created a new folder with a parser that can convert the literals to
values of 0 or 1.